backend/requirements.txt
google-cloud-aiplatform
fastapi
uvicorn
python-dotenv
pydantic
gTTS
sqlalchemy
pg8000
